Initial build
So this covers a hell of a lot and i was not committing sensibly,
bad adam.

It is essentially the initial build of the system which includes
most of the required functionality for managing the API resources
that represent a module and module version.

It needs alot more testing, and the functionality of actually storing
the module still needs to be added.